+++ apps/example/src/app/api/v2/route.ts
@@ -1,5 +1,7 @@
 import { docsRoute } from 'next-rest-framework';
 
+export const runtime = 'edge';
+
 export const { GET } = docsRoute({
   deniedPaths: ['/api/routes/third-party-endpoint'],
   openApiJsonPath: '/openapi.json'
