+++ apps/example/src/app/api/v2/rpc/[operationId]/route.ts
@@ -1,13 +1,13 @@
 import { createTodo, deleteTodo, getTodoById, getTodos } from '@/actions';
 import { rpcRoute } from 'next-rest-framework';
 
-const { POST, client } = rpcRoute({
+export const runtime = 'edge';
+
+export const { POST } = rpcRoute({
   getTodos,
   getTodoById,
   createTodo,
   deleteTodo
 });
 
-export type RpcClient = typeof client;
-
-export { POST };
+export type RpcClient = typeof POST.client;
