+++ apps/example/src/app/api/v2/third-party-endpoint/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from 'next/server';
 
+export const runtime = 'edge';
+
 // You can still write regular routes with Next REST Framework.
 export const GET = () => {
   return NextResponse.json('Server error', {
