+++ apps/example/src/pages/api/v1/rpc/[operationId].ts
