+++ docs/docs/intro.md
@@ -19,7 +19,7 @@ Next REST Framework is an open-source, opinionated, lightweight, easy-to-use set
 - Auto-generated and extensible `openapi.json` spec file from your business logic.
 - Auto-generated [Redoc](https://github.com/Redocly/redoc) and/or [SwaggerUI](https://swagger.io/tools/swagger-ui/) documentation frontend.
 - Works with Next.js [Middleware](https://nextjs.org/docs/app/building-your-application/routing/middleware) and other server-side libraries, like [NextAuth.js](#https://github.com/nextauthjs/next-auth).
-- Supports both Next.js [App Router](https://nextjs.org/docs/app/building-your-application/routing#the-app-router) and [Pages Router](https://nextjs.org/docs/pages/building-your-application/routing), even at the same time.
+- Supports both Next.js [app router](https://nextjs.org/docs/app/building-your-application/routing#the-app-router) and [pages router](https://nextjs.org/docs/pages/building-your-application/routing), even at the same time.
 - Fully customizable and compatible with any existing Next.js project.
 
 ### [Installation](#installation)
