+++ packages/next-rest-framework/package.json
@@ -19,7 +19,7 @@
     "dist"
   ],
   "main": "dist/index.js",
-  "module": "dist/index.esm.js",
+  "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
@@ -28,25 +28,26 @@
   },
   "scripts": {
     "lint": "tsc",
-    "test": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js",
+    "test": "jest",
     "test:watch": "jest --watch",
-    "build": "tsup src/** --format cjs,esm --dts"
+    "build": "tsup --dts"
   },
   "bin": {
-    "next-rest-framework": "./dist/cli.js"
+    "next-rest-framework": "./dist/cli/index.js"
   },
   "dependencies": {
+    "@esbuild-plugins/tsconfig-paths": "^0.1.2",
     "chalk": "4.1.2",
     "commander": "10.0.1",
+    "esbuild": "0.19.11",
     "lodash": "4.17.21",
     "prettier": "3.0.2",
-    "wait-on": "7.0.1",
+    "tiny-glob": "0.2.9",
     "zod-to-json-schema": "3.21.4"
   },
   "devDependencies": {
     "@types/jest": "29.5.4",
     "@types/lodash": "4.14.197",
-    "@types/wait-on": "5.3.1",
     "jest": "29.6.4",
     "node-mocks-http": "1.13.0",
     "openapi-types": "12.1.3",
