+++ packages/next-rest-framework/src/cli/constants.ts
@@ -1,0 +1,2 @@
+export const OPEN_API_VERSION = '3.0.1';
+export const NEXT_REST_FRAMEWORK_TEMP_FOLDER_NAME = '.next-rest-framework';
