+++ packages/next-rest-framework/src/constants.ts
@@ -1,21 +1,14 @@
-export const DEFAULT_MESSAGES = {
-  created: 'Created',
-  noContent: 'No content'
-};
-
 export const DEFAULT_ERRORS = {
   unexpectedError: 'An unknown error occurred, trying again might help.',
   methodNotAllowed: 'Method not allowed.',
   notFound: 'Not found.',
+  notImplemented: 'Not implemented.',
   invalidMediaType: 'Invalid media type.',
   operationNotAllowed: 'Operation not allowed.',
   invalidRequestBody: 'Invalid request body.',
   missingRequestBody: 'Missing request body.',
-  invalidQueryParameters: 'Invalid query parameters.',
-  handlerNotFound: 'Handler not found.'
+  invalidQueryParameters: 'Invalid query parameters.'
 };
-
-export const OPEN_API_VERSION = '3.0.1';
 
 export enum ValidMethod {
   GET = 'GET',
@@ -26,8 +19,6 @@ export enum ValidMethod {
   HEAD = 'HEAD',
   PATCH = 'PATCH'
 }
-
-export const NEXT_REST_FRAMEWORK_USER_AGENT = 'next-rest-framework';
 
 // Ignore: We don't want to use promises here to avoid making this an async function.
 // eslint-disable-next-line @typescript-eslint/no-var-requires
