+++ packages/next-rest-framework/src/global.d.ts
@@ -1,3 +1,4 @@
 declare module globalThis {
   var _nextRestFrameworkConfig;
+  var EdgeRuntime;
 }
