+++ packages/next-rest-framework/src/shared/config.ts
@@ -19,7 +19,6 @@ export const DEFAULT_CONFIG: Required<NextRestFrameworkConfig> = {
     }
   },
   openApiJsonPath: '/openapi.json',
-  autoGenerateOpenApiSpec: true,
   docsConfig: {
     provider: 'redoc',
     title: DEFAULT_TITLE,
