+++ packages/next-rest-framework/src/shared/index.ts
@@ -1,6 +1,7 @@
 export * from './config';
 export * from './docs';
 export * from './logging';
-export * from './open-api';
+export * from './paths';
 export * from './rpc-operation';
 export * from './schemas';
+export * from './utils';
