+++ packages/next-rest-framework/src/shared/logging.ts
@@ -1,37 +1,21 @@
 import chalk from 'chalk';
-import { type NextRestFrameworkConfig } from '../types';
-import { isEqualWith } from 'lodash';
 
-export const logInitInfo = ({
-  config,
-  baseUrl,
-  url
-}: {
-  config: Required<NextRestFrameworkConfig>;
-  baseUrl: string;
-  url: string;
-}) => {
-  const configsEqual = isEqualWith(global._nextRestFrameworkConfig, config);
+export const logPagesEdgeRuntimeErrorForRoute = (route: string) => {
+  console.error(
+    chalk.red(`---
+${route} is using Edge runtime in \`/pages\` folder that is not supported with \`apiRoute\`.
+Please use \`route\` instead: https://vercel.com/docs/functions/edge-functions/quickstart
+---`)
+  );
+};
 
-  const logReservedPaths = () => {
-    console.info(
-      chalk.yellowBright(`Docs: ${url}
-OpenAPI JSON: ${baseUrl}${config.openApiJsonPath}`)
-    );
-  };
-
-  if (!global._nextRestFrameworkConfig) {
-    global._nextRestFrameworkConfig = config;
-    console.info(chalk.green('Next REST Framework initialized! 🚀'));
-    logReservedPaths();
-  } else if (!configsEqual) {
-    console.info(
-      chalk.green('Next REST Framework config changed, re-initializing!')
-    );
-
-    global._nextRestFrameworkConfig = config;
-    logReservedPaths();
-  }
+export const logPagesEdgeRuntimeErrorForDocsRoute = (route: string) => {
+  console.error(
+    chalk.red(`---
+${route} is using Edge runtime in \`/pages\` folder that is not supported with \`docsApiRoute\`.
+Please use \`docsRoute\` instead: https://vercel.com/docs/functions/edge-functions/quickstart
+---`)
+  );
 };
 
 export const logNextRestFrameworkError = (error: unknown) => {
